Moving polyhedron gravity functions to basic_mathematics namespace
