+++ include/tudat/astro/gravitation/polyhedronGravityField.h
@@ -25,127 +25,13 @@
 
 #include "tudat/astro/gravitation/gravityFieldModel.h"
 #include "tudat/astro/basic_astro/polyhedronFuntions.h"
+#include "tudat/math/basic/polyhedron.h"
 
 namespace tudat
 {
 
 namespace gravitation
 {
-
-/*! Computes the coordinates of the polyhedron vertices with respect a the field point.
- *
- * Computes the coordinates of the polyhedron vertices with respect to the field point.
- * @param verticesCoordinatesRelativeToFieldPoint Matrix with coordinates of each vertex wrt field point (output).
- * @param bodyFixedPosition Body fixed position of field point (input).
- * @param verticesCoordinates Coordinates of polyehdron vertices (input).
- */
-void calculatePolyhedronVerticesCoordinatesRelativeToFieldPoint (
-        Eigen::MatrixXd& verticesCoordinatesRelativeToFieldPoint,
-        const Eigen::Vector3d& bodyFixedPosition,
-        const Eigen::MatrixXd& verticesCoordinates);
-
-/*! Calculates the per-facet factor of each polyhedron facet.
- *
- * Calculates the per-facet factor of each polyhedron facet, according to Eq. 27 of Werner and Scheeres (1997).
- * @param perFacetFactor Vector with the per-facet factor of each facet (output).
- * @param verticesCoordinatesRelativeToFieldPoint Matrix with coordinates of each vertex wrt field point (input).
- * @param verticesDefiningEachFacet Identification of the vertices defining each facet (0 indexed) (input)
- */
-void calculatePolyhedronPerFacetFactor (
-        Eigen::VectorXd& perFacetFactor,
-        const Eigen::MatrixXd& verticesCoordinatesRelativeToFieldPoint,
-        const Eigen::MatrixXi& verticesDefiningEachFacet);
-
-/*! Calculates the per-edge factor of each polyhedron edge.
- *
- * Calculates the per-edge factor of each polyhedron edge, according to Eq. 7 of Werner and Scheeres (1997).
- * @param perEdgeFactor Vector with the per-edge factor of each edge (output).
- * @param verticesCoordinatesRelativeToFieldPoint Matrix with coordinates of each vertex wrt field point (input).
- * @param verticesDefiningEachEdge Identification of the vertices defining each facet (0 indexed) (input)
- */
-void calculatePolyhedronPerEdgeFactor (
-        Eigen::VectorXd& perEdgeFactor,
-        const Eigen::MatrixXd& verticesCoordinatesRelativeToFieldPoint,
-        const Eigen::MatrixXi& verticesDefiningEachEdge);
-
-/*! Calculates the gravitational potential of a constant-density polyhedron.
- *
- * Calculates the gravitational potential of a constant-density polyhedron, according to Eq. 10 of Werner and Scheeres
- * (1997).
- * @param gravitationalConstantTimesDensity Product of the gravitational constant and density.
- * @param verticesCoordinatesRelativeToFieldPoint Matrix with coordinates of each vertex wrt field point.
- * @param verticesDefiningEachFacet Identification of the vertices defining each facet (0 indexed).
- * @param verticesDefiningEachEdge Identification of the vertices defining each edge (0 indexed).
- * @param facetDyads Vector containing facet dyads.
- * @param edgeDyads Vector containing edge dyads.
- * @param perFacetFactor Vector containing per-facet factors.
- * @param perEdgeFactor Vector containing per-edge factors.
- * @return Gravitational potential.
- */
-double calculatePolyhedronGravitationalPotential(
-        const double gravitationalConstantTimesDensity,
-        const Eigen::MatrixXd& verticesCoordinatesRelativeToFieldPoint,
-        const Eigen::MatrixXi& verticesDefiningEachFacet,
-        const Eigen::MatrixXi& verticesDefiningEachEdge,
-        const std::vector< Eigen::MatrixXd >& facetDyads,
-        const std::vector< Eigen::MatrixXd >& edgeDyads,
-        const Eigen::VectorXd& perFacetFactor,
-        const Eigen::VectorXd& perEdgeFactor );
-
-/*! Calculates the gradient of the potential of a constant-density polyhedron.
- *
- * Calculates the gradient of the potential of a constant-density polyhedron, according to Eq. 15 of Werner and Scheeres
- * (1997).
- * @param gravitationalConstantTimesDensity Product of the gravitational constant and density.
- * @param verticesCoordinatesRelativeToFieldPoint Matrix with coordinates of each vertex wrt field point.
- * @param verticesDefiningEachFacet Identification of the vertices defining each facet (0 indexed).
- * @param verticesDefiningEachEdge Identification of the vertices defining each edge (0 indexed).
- * @param facetDyads Vector containing facet dyads.
- * @param edgeDyads Vector containing edge dyads.
- * @param perFacetFactor Vector containing per-facet factors.
- * @param perEdgeFactor Vector containing per-edge factors.
- * @return Gradient of gravitational potential.
- */
-Eigen::Vector3d calculatePolyhedronGradientOfGravitationalPotential(
-        const double gravitationalConstantTimesDensity,
-        const Eigen::MatrixXd& verticesCoordinatesRelativeToFieldPoint,
-        const Eigen::MatrixXi& verticesDefiningEachFacet,
-        const Eigen::MatrixXi& verticesDefiningEachEdge,
-        const std::vector< Eigen::MatrixXd >& facetDyads,
-        const std::vector< Eigen::MatrixXd >& edgeDyads,
-        const Eigen::VectorXd& perFacetFactor,
-        const Eigen::VectorXd& perEdgeFactor);
-
-/*! Calculates the Hessian matrix of the potential of a constant-density polyhedron.
- *
- * Calculates the Hessian matrix of the potential of a constant-density polyhedron, according to Eq. 16 of Werner and
- * Scheeres (1997).
- * @param gravitationalConstantTimesDensity Product of the gravitational constant and density.
- * @param facetDyads Vector containing facet dyads.
- * @param edgeDyads Vector containing edge dyads.
- * @param perFacetFactor Vector containing per-facet factors.
- * @param perEdgeFactor Vector containing per-edge factors.
- * @return Hessian matrix of the potential.
- */
-Eigen::Matrix3d calculatePolyhedronHessianOfGravitationalPotential(
-        const double gravitationalConstantTimesDensity,
-        const std::vector< Eigen::MatrixXd >& facetDyads,
-        const std::vector< Eigen::MatrixXd >& edgeDyads,
-        const Eigen::VectorXd& perFacetFactor,
-        const Eigen::VectorXd& perEdgeFactor);
-
-/*! Calculate the laplacian of the gravitational potential of a constant-density polyhedron.
- *
- * Calculate the laplacian of the gravitational potential of a constant-density polyhedron, according to Eq. 17 of
- * Werner and Scheeres (1997).
- * @param gravitationalConstantTimesDensity Product of the gravitational constant and density.
- * @param perFacetFactor Vector containing per-facet factors.
- * @return Laplacian of the gravitational potential
- */
-double calculatePolyhedronLaplacianOfGravitationalPotential(
-        const double gravitationalConstantTimesDensity,
-        const Eigen::VectorXd& perFacetFactor);
-
 
 //! Cache object in which variables that are required for the computation of polyhedron gravity field are stored.
 class PolyhedronGravityCache
@@ -232,7 +118,7 @@ private:
 };
 
 
-//! Class to represent the gravity field of a constant density polyhedron
+//! Class to represent the gravity field of a constant density polyhedron.
 class PolyhedronGravityField: public GravityFieldModel
 {
 public:
@@ -263,7 +149,7 @@ public:
         fixedReferenceFrame_( fixedReferenceFrame )
     {
         // Check if provided arguments are valid
-        polyhedron_utilities::checkValidityOfPolyhedronSettings(verticesCoordinates, verticesDefiningEachFacet);
+        basic_mathematics::checkValidityOfPolyhedronSettings(verticesCoordinates, verticesDefiningEachFacet);
 
         // Compute edges in polyhedron
         computeVerticesAndFacetsDefiningEachEdge();
@@ -289,7 +175,7 @@ public:
     {
         polyhedronGravityCache_->update(bodyFixedPosition);
 
-        return calculatePolyhedronGravitationalPotential(
+        return basic_mathematics::calculatePolyhedronGravitationalPotential(
                 gravitationalParameter_ / volume_,
                 polyhedronGravityCache_->getVerticesCoordinatesRelativeToFieldPoint(),
                 verticesDefiningEachFacet_,
@@ -310,7 +196,7 @@ public:
     {
         polyhedronGravityCache_->update(bodyFixedPosition);
 
-        return calculatePolyhedronGradientOfGravitationalPotential(
+        return basic_mathematics::calculatePolyhedronGradientOfGravitationalPotential(
                 gravitationalParameter_ / volume_,
                 polyhedronGravityCache_->getVerticesCoordinatesRelativeToFieldPoint(),
                 verticesDefiningEachFacet_,
@@ -331,7 +217,7 @@ public:
     {
         polyhedronGravityCache_->update(bodyFixedPosition);
 
-        return calculatePolyhedronHessianOfGravitationalPotential(
+        return basic_mathematics::calculatePolyhedronHessianOfGravitationalPotential(
                 gravitationalParameter_ / volume_,
                 facetDyads_,
                 edgeDyads_,
@@ -349,7 +235,7 @@ public:
     {
         polyhedronGravityCache_->update(bodyFixedPosition);
 
-        return calculatePolyhedronLaplacianOfGravitationalPotential(
+        return basic_mathematics::calculatePolyhedronLaplacianOfGravitationalPotential(
                 gravitationalParameter_ / volume_,
                 polyhedronGravityCache_->getPerFacetFactor() );
     }
