+++ src/math/basic/CMakeLists.txt
@@ -15,6 +15,7 @@ set(basic_mathematics_SOURCES
         "nearestNeighbourSearch.cpp"
         "numericalDerivative.cpp"
         "sphericalHarmonics.cpp"
+        "polyhedron.cpp"
         "rotationAboutArbitraryAxis.cpp"
         "basicMathematicsFunctions.cpp"
         "coordinateConversions.cpp"
@@ -35,6 +36,7 @@ set(basic_mathematics_HEADERS
         "nearestNeighbourSearch.h"
         "numericalDerivative.h"
         "sphericalHarmonics.h"
+        "polyhedron.h"
         "rotationAboutArbitraryAxis.h"
         "basicMathematicsFunctions.h"
         "coordinateConversions.h"
